@@ -20,29 +20,50 @@
 
 namespace Phalcon\Db\Dialect;
 
+use Phalcon\Db\Exception;
+
 /**
- * Phalcon\Db\Adapter\Dialect\MysqlExtended
+ * Phalcon\Db\Dialect\MysqlExtended
  *
- * Every query executed via this adapter is automatically cached
+ * Generates database specific SQL for the MySQL RDBMS.
+ *
+ * This is an extended MySQL dialect that introduces workarounds for some common MySQL-only functions like
+ * search based on FULLTEXT indexes and operations with date intervals.
+ *
+ * <code>
+ * use Phalcon\Db\Adapter\Pdo\Mysql;
+ * use Phalcon\Db\Adapter\Pdo\MysqlExtended;
+ *
+ * $connection = new Mysql([
+ *     'host'         => 'localhost',
+ *     'username'     => 'root',
+ *     'password'     => 'secret',
+ *     'dbname'       => 'enigma',
+ *     'dialectClass' => MysqlExtended::class
+ * ]);
+ * </code>
+ *
+ * @package Phalcon\Db\Dialect
  */
 class MysqlExtended extends Mysql
 {
     /**
-     * Transforms an intermediate representation for a expression into a database system valid expression
+     * Transforms an intermediate representation for a expression into a database system valid expression.
      *
-     * @param array  $expression
+     * @param array $expression
      * @param string $escapeChar
-     *
+     * @param mixed $bindCounts
      * @return string
-     * @throws \Exception
+     *
+     * @throws Exception
      */
     public function getSqlExpression(array $expression, $escapeChar = null, $bindCounts = null)
     {
         if ($expression["type"] == 'functionCall') {
-            switch ($expression["name"]) {
+            switch (strtoupper($expression["name"])) {
                 case 'DATE_INTERVAL':
                     if (count($expression["arguments"]) != 2) {
-                        throw new \Exception('DATE_INTERVAL requires 2 parameters');
+                        throw new Exception('DATE_INTERVAL requires 2 parameters');
                     }
 
                     switch ($expression["arguments"][1]['value']) {
@@ -71,10 +92,10 @@ class MysqlExtended extends Mysql
 
                 case 'FULLTEXT_MATCH':
                     if (count($expression["arguments"]) < 2) {
-                        throw new \Exception('FULLTEXT_MATCH requires 2 parameters');
+                        throw new Exception('FULLTEXT_MATCH requires 2 parameters');
                     }
 
-                    $arguments = array();
+                    $arguments = [];
                     $length = count($expression["arguments"]) - 1;
                     for ($i = 0; $i < $length; $i++) {
                         $arguments[] = $this->getSqlExpression($expression["arguments"][$i]);
@@ -82,13 +103,14 @@ class MysqlExtended extends Mysql
 
                     return 'MATCH(' . join(', ', $arguments) . ') AGAINST (' .
                     $this->getSqlExpression($expression["arguments"][$length]) . ')';
+                    break;
 
                 case 'FULLTEXT_MATCH_BMODE':
                     if (count($expression["arguments"]) < 2) {
-                        throw new \Exception('FULLTEXT_MATCH requires 2 parameters');
+                        throw new Exception('FULLTEXT_MATCH requires 2 parameters');
                     }
 
-                    $arguments = array();
+                    $arguments = [];
                     $length = count($expression["arguments"]) - 1;
                     for ($i = 0; $i < $length; $i++) {
                         $arguments[] = $this->getSqlExpression($expression["arguments"][$i]);
@@ -96,9 +118,19 @@ class MysqlExtended extends Mysql
 
                     return 'MATCH(' . join(', ', $arguments) . ') AGAINST (' .
                     $this->getSqlExpression($expression["arguments"][$length]) . ' IN BOOLEAN MODE)';
+                    break;
+
+                case 'REGEXP':
+                    if (count($expression['arguments']) != 2) {
+                        throw new Exception('REGEXP requires 2 parameters');
+                    }
+
+                    return $this->getSqlExpression($expression['arguments'][0]) .
+                    ' REGEXP (' . $this->getSqlExpression($expression['arguments'][1]) . ')';
+                    break;
             }
         }
 
-        return parent::getSqlExpression($expression, $escapeChar);
+        return parent::getSqlExpression($expression, $escapeChar, $bindCounts);
     }
 }
