@@ -122,7 +122,7 @@ class Header implements \Countable
             return false;
         }
 
-        $status = array();
+        $status = [];
         if (preg_match('%^HTTP/(\d(?:\.\d)?)\s+(\d{3})\s?+(.+)?$%i', $content[0], $status)) {
             $this->status = array_shift($content);
             $this->version = $status[1];
@@ -149,7 +149,7 @@ class Header implements \Countable
      */
     public function build($flags = 0)
     {
-        $lines = array();
+        $lines = [];
         if (($flags & self::BUILD_STATUS) && StatusCode::message($this->statusCode)) {
             $lines[] = 'HTTP/' . $this->version . ' ' .
                 $this->statusCode . ' ' .
