@@ -54,12 +54,12 @@ class Stream extends Request
 
     private function initOptions()
     {
-        $this->setOptions(array(
+        $this->setOptions([
             'user_agent'      => 'Phalcon HTTP/' . self::VERSION . ' (Stream)',
             'follow_location' => 1,
             'max_redirects'   => 20,
             'timeout'         => 30
-        ));
+        ]);
     }
 
     public function setOption($option, $value)
@@ -69,7 +69,7 @@ class Stream extends Request
 
     public function setOptions($options)
     {
-        return stream_context_set_option($this->context, array('http' => $options));
+        return stream_context_set_option($this->context, ['http' => $options]);
     }
 
     public function setTimeout($timeout)
@@ -82,13 +82,13 @@ class Stream extends Request
         throw new HttpException($errstr, $errno);
     }
 
-    private function send($uri)
+    protected function send($uri)
     {
         if (count($this->header) > 0) {
             $this->setOption('header', $this->header->build(Header::BUILD_FIELDS));
         }
 
-        set_error_handler(array($this, 'errorHandler'));
+        set_error_handler([$this, 'errorHandler']);
         $content = file_get_contents($uri->build(), false, $this->context);
         restore_error_handler();
 
@@ -99,7 +99,7 @@ class Stream extends Request
         return $response;
     }
 
-    private function initPostFields($params)
+    protected function initPostFields($params)
     {
         if (!empty($params) && is_array($params)) {
             $this->header->set('Content-Type', 'application/x-www-form-urlencoded');
@@ -109,11 +109,11 @@ class Stream extends Request
 
     public function setProxy($host, $port = 8080, $user = null, $pass = null)
     {
-        $uri = new Uri(array(
+        $uri = new Uri([
             'scheme' => 'tcp',
             'host'   => $host,
             'port'   => $port
-        ));
+        ]);
 
         if (!empty($user)) {
             $uri->user = $user;
@@ -125,7 +125,7 @@ class Stream extends Request
         $this->setOption('proxy', $uri->build());
     }
 
-    public function get($uri, $params = array())
+    public function get($uri, $params = [])
     {
         $uri = $this->resolveUri($uri);
 
@@ -133,17 +133,17 @@ class Stream extends Request
             $uri->extendQuery($params);
         }
 
-        $this->setOptions(array(
+        $this->setOptions([
             'method'  => Method::GET,
             'content' => ''
-        ));
+        ]);
 
         $this->header->remove('Content-Type');
 
         return $this->send($uri);
     }
 
-    public function head($uri, $params = array())
+    public function head($uri, $params = [])
     {
         $uri = $this->resolveUri($uri);
 
@@ -151,17 +151,17 @@ class Stream extends Request
             $uri->extendQuery($params);
         }
 
-        $this->setOptions(array(
+        $this->setOptions([
             'method'  => Method::HEAD,
             'content' => ''
-        ));
+        ]);
 
         $this->header->remove('Content-Type');
 
         return $this->send($uri);
     }
 
-    public function delete($uri, $params = array())
+    public function delete($uri, $params = [])
     {
         $uri = $this->resolveUri($uri);
 
@@ -169,17 +169,17 @@ class Stream extends Request
             $uri->extendQuery($params);
         }
 
-        $this->setOptions(array(
+        $this->setOptions([
             'method'  => Method::DELETE,
             'content' => ''
-        ));
+        ]);
 
         $this->header->remove('Content-Type');
 
         return $this->send($uri);
     }
 
-    public function post($uri, $params = array())
+    public function post($uri, $params = [])
     {
         $this->setOption('method', Method::POST);
 
@@ -188,9 +188,18 @@ class Stream extends Request
         return $this->send($this->resolveUri($uri));
     }
 
-    public function put($uri, $params = array())
+    public function put($uri, $params = [])
     {
         $this->setOption('method', Method::PUT);
+
+        $this->initPostFields($params);
+
+        return $this->send($this->resolveUri($uri));
+    }
+
+    public function patch($uri, $params = [])
+    {
+        $this->setOption('method', Method::PATCH);
 
         $this->initPostFields($params);
 
