@@ -76,7 +76,7 @@ class Blameable extends Behavior implements BehaviorInterface
         $audit    = $this->createAudit('C', $model);
         $metaData = $model->getModelsMetaData();
         $fields   = $metaData->getAttributes($model);
-        $details  = array();
+        $details  = [];
 
         foreach ($fields as $field) {
             $auditDetail = new AuditDetail();
@@ -112,7 +112,7 @@ class Blameable extends Behavior implements BehaviorInterface
         //Date the model had before modifications
         $originalData = $model->getSnapshotData();
 
-        $details = array();
+        $details = [];
         foreach ($changedFields as $field) {
             $auditDetail = new AuditDetail();
             $auditDetail->field_name = $field;
