@@ -28,8 +28,8 @@ class Twig extends Engine implements EngineInterface
     public function __construct(
         ViewBaseInterface $view,
         DiInterface $di = null,
-        $options = array(),
-        $userFunctions = array()
+        $options = [],
+        $userFunctions = []
     ) {
         $loader     = new \Twig_Loader_Filesystem($view->getViewsDir());
         $this->twig = new Twig\Environment($di, $loader, $options);
@@ -47,13 +47,11 @@ class Twig extends Engine implements EngineInterface
      * @param \Phalcon\DiInterface       $di
      * @param array                      $userFunctions
      */
-    protected function registryFunctions($view, DiInterface $di, $userFunctions = array())
+    protected function registryFunctions($view, DiInterface $di, $userFunctions = [])
     {
-        $options = array(
-            'is_safe' => array('html')
-        );
+        $options = ['is_safe' => ['html']];
 
-        $functions = array(
+        $functions = [
             new \Twig_SimpleFunction('content', function () use ($view) {
                 return $view->getContent();
             }, $options),
@@ -84,16 +82,16 @@ class Twig extends Engine implements EngineInterface
             new \Twig_SimpleFunction('submitButton', function ($parameters) {
                 return \Phalcon\Tag::submitButton($parameters);
             }, $options),
-            new \Twig_SimpleFunction('selectStatic', function ($parameters, $data = array()) {
+            new \Twig_SimpleFunction('selectStatic', function ($parameters, $data = []) {
                 return \Phalcon\Tag::selectStatic($parameters, $data);
             }, $options),
-            new \Twig_SimpleFunction('select', function ($parameters, $data = array()) {
+            new \Twig_SimpleFunction('select', function ($parameters, $data = []) {
                 return \Phalcon\Tag::select($parameters, $data);
             }, $options),
             new \Twig_SimpleFunction('textArea', function ($parameters) {
                 return \Phalcon\Tag::textArea($parameters);
             }, $options),
-            new \Twig_SimpleFunction('form', function ($parameters = array()) {
+            new \Twig_SimpleFunction('form', function ($parameters = []) {
                 return \Phalcon\Tag::form($parameters);
             }, $options),
             new \Twig_SimpleFunction('endForm', function () {
@@ -126,7 +124,7 @@ class Twig extends Engine implements EngineInterface
             new \Twig_SimpleFunction('url', function ($route) use ($di) {
                 return $di->get("url")->get($route);
             }, $options)
-        );
+        ];
 
         if (!empty($userFunctions)) {
             $functions = array_merge($functions, $userFunctions);
