@@ -70,6 +70,12 @@ class Database extends Adapter
     protected $rolesInherits;
 
     /**
+     * Default action for no arguments is allow
+     * @var int
+     */
+    protected $noArgumentsDefaultAction = Acl::ALLOW;
+
+    /**
      * Class constructor.
      *
      * @param  array $options Adapter config
@@ -98,8 +104,10 @@ class Database extends Adapter
      * {@inheritdoc}
      *
      * Example:
-     * <code>$acl->addRole(new Phalcon\Acl\Role('administrator'), 'consultor');</code>
-     * <code>$acl->addRole('administrator', 'consultor');</code>
+     * <code>
+     * $acl->addRole(new Phalcon\Acl\Role('administrator'), 'consultor');
+     * $acl->addRole('administrator', 'consultor');
+     * </code>
      *
      * @param  \Phalcon\Acl\Role|string $role
      * @param  string                   $accessInherits
@@ -342,8 +350,9 @@ class Database extends Adapter
      * @param string       $roleName
      * @param string       $resourceName
      * @param array|string $access
+     * @param mixed $func
      */
-    public function allow($roleName, $resourceName, $access)
+    public function allow($roleName, $resourceName, $access, $func = null)
     {
         $this->allowOrDeny($roleName, $resourceName, $access, Acl::ALLOW);
     }
@@ -366,9 +375,10 @@ class Database extends Adapter
      * @param  string       $roleName
      * @param  string       $resourceName
      * @param  array|string $access
+     * @param  mixed $func
      * @return boolean
      */
-    public function deny($roleName, $resourceName, $access)
+    public function deny($roleName, $resourceName, $access, $func = null)
     {
         $this->allowOrDeny($roleName, $resourceName, $access, Acl::DENY);
     }
@@ -386,10 +396,10 @@ class Database extends Adapter
      * @param string $role
      * @param string $resource
      * @param string $access
-     *
+     * @param array  $parameters
      * @return bool
      */
-    public function isAllowed($role, $resource, $access)
+    public function isAllowed($role, $resource, $access, array $parameters = null)
     {
         $sql = implode(' ', [
             "SELECT " . $this->connection->escapeIdentifier('allowed') . " FROM {$this->accessList} AS a",
@@ -423,6 +433,28 @@ class Database extends Adapter
          */
 
         return $this->_defaultAccess;
+    }
+
+    /**
+     * Returns the default ACL access level for no arguments provided
+     * in isAllowed action if there exists func for accessKey
+     *
+     * @return int
+     */
+    public function getNoArgumentsDefaultAction()
+    {
+        return $this->noArgumentsDefaultAction;
+    }
+
+    /**
+     * Sets the default access level for no arguments provided
+     * in isAllowed action if there exists func for accessKey
+     *
+     * @param int $defaultAccess Phalcon\Acl::ALLOW or Phalcon\Acl::DENY
+     */
+    public function setNoArgumentsDefaultAction($defaultAccess)
+    {
+        $this->noArgumentsDefaultAction = intval($defaultAccess);
     }
 
     /**
