@@ -91,7 +91,7 @@ MSG;
 
         $this->subject = $from;
         $this->subjectClassName = $className;
-        $this->eagerLoads = ($from === null || empty($arguments)) ? array() : static::parseArguments($arguments);
+        $this->eagerLoads = ($from === null || empty($arguments)) ? [] : static::parseArguments($arguments);
     }
 
     /**
