@@ -38,10 +38,16 @@ use Phalcon\Validation;
  *         'table'   => 'users',
  *         'column'  => 'login',
  *         'message' => 'already taken',
+ *         'exclude' => [
+ *             'column' => 'id',
+ *             'value' => 1 // Some ID to exclude
+ *         ],
  *     ],
  *     $di->get('db');
  * );
  * </code>
+ *
+ * Exclude option is optional.
  *
  * If second parameter will be null (omitted) than validator will try to get database
  * connection from default DI instance with \Phalcon\Di::getDefault()->get('db');
@@ -62,11 +68,11 @@ class Uniqueness extends Validator
      * @param  DbConnection  $db
      * @throws ValidationException
      */
-    public function __construct(array $options = [], $db = null)
+    public function __construct(array $options = [], DbConnection $db = null)
     {
         parent::__construct($options);
 
-        if (null === $db) {
+        if (!$db) {
             // try to get db instance from default Dependency Injection
             $di = Di::getDefault();
 
@@ -75,16 +81,28 @@ class Uniqueness extends Validator
             }
         }
 
-        if (!($db instanceof DbConnection)) {
-            throw new ValidationException('Validator Uniquness require connection to database');
+        if (!$db instanceof DbConnection) {
+            throw new ValidationException('Validator Uniqueness require connection to database');
         }
 
-        if (false === $this->hasOption('table')) {
+        if (!$this->hasOption('table')) {
             throw new ValidationException('Validator require table option to be set');
         }
 
-        if (false === $this->hasOption('column')) {
+        if (!$this->hasOption('column')) {
             throw new ValidationException('Validator require column option to be set');
+        }
+
+        if ($this->hasOption('exclude')) {
+            $exclude = $this->getOption('exclude');
+
+            if (!isset($exclude['column']) || empty($exclude['column'])) {
+                throw new ValidationException('Validator with "exclude" option require column option to be set');
+            }
+
+            if (!isset($exclude['value']) || empty($exclude['value'])) {
+                throw new ValidationException('Validator with "exclude" option require value option to be set');
+            }
         }
 
         $this->db = $db;
@@ -99,22 +117,31 @@ class Uniqueness extends Validator
      */
     public function validate(Validation $validator, $attribute)
     {
-        $table = $this->getOption('table');
-        $column = $this->getOption('column');
+        $table = $this->db->escapeIdentifier($this->getOption('table'));
+        $column = $this->db->escapeIdentifier($this->getOption('column'));
 
-        $result = $this->db->fetchOne(
-            sprintf('SELECT COUNT(*) as count FROM %s WHERE %s = ?', $table, $column),
-            Db::FETCH_ASSOC,
-            [$validator->getValue($attribute)]
-        );
+        if ($this->hasOption('exclude')) {
+            $exclude = $this->getOption('exclude');
+            $result = $this->db->fetchOne(
+                sprintf(
+                    'SELECT COUNT(*) AS count FROM %s WHERE %s = ? AND %s != ?',
+                    $table,
+                    $column,
+                    $this->db->escapeIdentifier($exclude['column'])
+                ),
+                Db::FETCH_ASSOC,
+                [$validator->getValue($attribute), $exclude['value']]
+            );
+        } else {
+            $result = $this->db->fetchOne(
+                sprintf('SELECT COUNT(*) AS count FROM %s WHERE %s = ?', $table, $column),
+                Db::FETCH_ASSOC,
+                [$validator->getValue($attribute)]
+            );
+        }
 
         if ($result['count']) {
-            $message = $this->getOption('message');
-
-            if (null === $message) {
-                $message = 'Already taken. Choose another!';
-            }
-
+            $message = $this->getOption('message', 'Already taken. Choose another!');
             $validator->appendMessage(new Message($message, $attribute, 'Uniqueness'));
 
             return false;
