@@ -13,7 +13,7 @@
   | obtain it through the world-wide-web, please send an email             |
   | to license@phalconphp.com so we can send you a copy immediately.       |
   +------------------------------------------------------------------------+
-  | Authors: David Hubner <david.hubner@gmail.com>                         |
+  | Authors: David Hubner <david.hubner@gmail.com>                             |
   +------------------------------------------------------------------------+
  */
 
