@@ -42,6 +42,9 @@ class MemcachedTest extends Test
      */
     protected function _before()
     {
+        if (!extension_loaded('memcached')) {
+            $this->markTestSkipped('memcached extension not loaded');
+        }
     }
 
     /**
@@ -52,21 +55,12 @@ class MemcachedTest extends Test
     }
 
     /**
-     * @expectedException        \Phalcon\Mvc\Model\Exception
+     * @expectedException        \Phalcon\Annotations\Exception
      * @expectedExceptionMessage No host given in options
      */
     public function testShouldCatchExceptionWhenNoHostGivenInOptions()
     {
         new Memcached(['lifetime' => 23, 'prefix' => '']);
-    }
-
-    /**
-     * @expectedException        \Phalcon\Mvc\Model\Exception
-     * @expectedExceptionMessage No configuration given
-     */
-    public function testShouldCatchExceptionWhenNoConfigurationGiven()
-    {
-        new Memcached(1);
     }
 
     public function testHasDefaultPort()
@@ -86,7 +80,6 @@ class MemcachedTest extends Test
 
     /**
      * @dataProvider providerReadWrite
-     * @requires extension memcached
      * @param string $key
      * @param mixed $data
      */
@@ -100,7 +93,6 @@ class MemcachedTest extends Test
 
     /**
      * @dataProvider providerReadWrite
-     * @requires extension memcached
      * @param string $key
      * @param mixed $data
      */
@@ -112,9 +104,6 @@ class MemcachedTest extends Test
         $this->assertEquals($data, $object->read($key));
     }
 
-    /**
-     * @requires extension memcached
-     */
     public function testShouldGetCacheBackendThroughGetter()
     {
         $object = new Memcached(['host' => TEST_MC_HOST]);
@@ -124,9 +113,6 @@ class MemcachedTest extends Test
         $this->assertInstanceOf(self::LIBMEMCACHED_CLASS, $reflectedMethod->invoke($object));
     }
 
-    /**
-     * @requires extension memcached
-     */
     public function testShouldGetCacheBackendThroughReflectionSetter()
     {
         $object = new Memcached(['host' => TEST_MC_HOST]);
@@ -142,7 +128,7 @@ class MemcachedTest extends Test
     }
 
     /**
-     * @dataProvider providerReadWrite
+     * @dataProvider providerKey
      * @param mixed $key
      */
     public function testShouldPrepareKey($key)
