@@ -1,2 +1,3 @@
 <?php
-// Here you can initialize variables that will be available to your tests
+
+Codeception\Specify\Config::setDeepClone(false);
